feat: Add translation functionality and endpoints
The code changes in this commit add new functionality for translation. It includes the implementation of translation endpoints for translating text from English to Chinese and vice versa. This will allow users to easily translate text using the application.

Note: The commit message has been generated based on the provided code changes and recent user commits.